feat: NumPy 2.4.2 alignment investigation with bug fixes
This commit addresses findings from a comprehensive NumPy alignment
investigation, achieving 92% compatibility (84/91 tests pass).

## Bug Fixes

### Task #84: Empty array max/min throws ArgumentException
- `np.max([])` and `np.min([])` now throw ArgumentException
- Matches NumPy's "zero-size array to reduction operation" ValueError
- Files: Default.Reduction.AMax.cs, Default.Reduction.AMin.cs

### Task #85: argmax/argmin return Int64 (not Int32)
- Changed return type from int to long to match NumPy
- Updated IL kernel wrapper to use Convert.ToInt64() for boxed results
- Fixed test files expecting int to use long
- Files: Default.Reduction.ArgMax/Min.cs, NDArray.argmax/argmin.cs,
         DefaultEngine.ReductionOp.cs, np.argmax.cs

## New Features

### Task #86: np.nanmean, np.nanstd, np.nanvar
- Implements NaN-ignoring statistical functions
- Full axis support with keepdims parameter
- ddof parameter for std/var (delta degrees of freedom)
- Files: np.nanmean.cs, np.nanstd.cs, np.nanvar.cs

## Test Coverage

- TypePromotionTests.cs: 34 tests for NEP50 type promotion rules
- EmptyArrayReductionTests.cs: Empty array edge case validation
- Updated ArgMaxNaNTests.cs, NDArray.AMax.Test.cs for long return type

## Documentation

- docs/NUMPY_ALIGNMENT_INVESTIGATION.md: Comprehensive NumPy 2.4.2
  behavior reference and test results by category

Investigation tasks completed: #80-#86 (7 tasks)
Overall alignment: 92% (84/91 tests pass)

+++ src/NumSharp.Core/Backends/Default/Math/Reduction/Default.Reduction.ArgMin.cs
@@ -57,9 +57,9 @@ namespace NumSharp.Backends
                 if (keepdims)
                 {
                     // Keep the axis but reduce to size 1 (it's already 1)
-                    return np.zeros(shape.dimensions, NPTypeCode.Int32);
+                    return np.zeros(shape.dimensions, NPTypeCode.Int64);
                 }
-                return np.squeeze_fast(np.zeros(shape.dimensions, NPTypeCode.Int32), axis);
+                return np.squeeze_fast(np.zeros(shape.dimensions, NPTypeCode.Int64), axis);
             }
 
             //handle keepdims - prepare output shape
@@ -81,7 +81,7 @@ namespace NumSharp.Backends
             }
 
             //prepare ret
-            var ret = new NDArray(NPTypeCode.Int32, axisedShape, false);
+            var ret = new NDArray(NPTypeCode.Int64, axisedShape, false);
             var iterAxis = new NDCoordinatesAxisIncrementor(ref shape, axis);
             var iterRet = new ValueCoordinatesIncrementor(ref axisedShape);
             var iterIndex = iterRet.Index;
@@ -114,7 +114,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
                     break;
                 }
@@ -153,7 +153,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(minAt, iterIndex);
+                        ret.SetInt64(minAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -181,7 +181,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -209,7 +209,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -237,7 +237,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -265,7 +265,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -293,7 +293,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -321,7 +321,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -349,7 +349,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -377,7 +377,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -406,7 +406,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -435,7 +435,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
@@ -463,7 +463,7 @@ namespace NumSharp.Backends
                             idx++;
                         }
 
-                        ret.SetInt32(maxAt, iterIndex);
+                        ret.SetInt64(maxAt, iterIndex);
                     } while (iterAxis.Next() != null && iterRet.Next() != null);
 
                     break;
